+++ tests/test_python3.py
@@ -7,6 +7,9 @@ Test module for ``python3.sif`` build
 import os
 import subprocess
 
+# https://github.com/comorment/containers/issues/267:
+os.environ.update({'SINGULARITYENV_LD_LIBRARY_PATH': 
+                   '/usr/local/lib:$LD_LIBRARY_PATH'})
 
 pth = os.path.join('singularity', 'python3.sif')
 
@@ -71,3 +74,9 @@ def test_python3_packages():
         call = f'singularity run {pth} python -c "import {pkg}"'
         out = subprocess.run(call.split(' '))
         assert out.returncode == 0
+
+def test_python3_import_pandas_scipy_stats():
+    pwd = os.getcwd()
+    call = f'singularity run --home={pwd} {pth} python -c "import pandas as pd; from scipy import *"'
+    out = subprocess.run(call.split(' '))
+    assert out.returncode == 0