+++ CHANGELOG.md
@@ -51,6 +51,7 @@ If MD5 sum is not listed for a certain release then it means that the container 
 ### Fixed
 
 * Fixed broken unit test `tests/test_gwas.py::test_gwas_metal` with Apptainer "sandbox" mode
+* Workaround for pandas import before scipy in python codes via `export LD_LIBRARY_PATH=/usr/local/lib:$LD_LIBRARY_PATH`
 * Fixed brittle tests if `TMPDIR` is not `/tmp`
 
 ### Removed